+++ lib/auth/login_page.dart
@@ -3,7 +3,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:projectloner/auth/ForgotPassword.dart';
+import 'package:projectloner/auth/forgot_password.dart';
 
 class LoginPage extends StatefulWidget {
   final VoidCallback showRegisterPage;
