+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'views/home_view.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:projectloner/auth/CheckLogin.dart';
+import 'package:projectloner/auth/check_login.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:projectloner/blocs/swipe/swipe_bloc.dart';
 import 'package:projectloner/views/onboarding_screen.dart';
