+++ lib/views/nav_bar.dart
@@ -4,7 +4,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import "package:flutter/material.dart";
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'home_view.dart';
-import 'package:projectloner/auth/AuthPage.dart';
+import 'package:projectloner/auth/auth_page.dart';
 import 'package:projectloner/matching/matching_screen.dart';
 import 'package:projectloner/profile/profile.dart';
